requirements.txt
asgiref==3.5.2
autopep8==2.0.0
distlib==0.3.6
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
drf_yasg==1.21.4
filelock==3.8.2
platformdirs==2.5.4
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
virtualenv==20.17.1
